League support for backend

+++ src/main/kotlin/com/epifoos/domain/calculation/MatchResult.kt
@@ -1,11 +1,11 @@
 package com.epifoos.domain.calculation
 
-enum class MatchResult(val weight: Float) {
-    WIN(4.0F),
-    WIN_ON_SCORE(3.0F),
-    DRAW(2.0F),
-    LOSE_ON_SCORE(1.0F),
-    LOSE(0.0F);
+enum class MatchResult(val weight: Double) {
+    WIN(4.0),
+    WIN_ON_SCORE(3.0),
+    DRAW(2.0),
+    LOSE_ON_SCORE(1.0),
+    LOSE(0.0);
 
     companion object {
         fun max(): MatchResult = WIN
